refactor to remove mutual deps
