+++ src/comms/profile.rs
@@ -3,8 +3,10 @@ use ethers::types::Address;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    avatar::AvatarWireFormat, dcl_component::proto_components::kernel::comms::rfc4,
-    scene_runner::PrimaryUser, util::AsH160, AppConfig,
+    common::PrimaryUser,
+    dcl_component::proto_components::{common::Color3, kernel::comms::rfc4},
+    util::AsH160,
+    AppConfig,
 };
 
 use super::{
@@ -236,6 +238,36 @@ pub fn process_profile_events(
     }
 
     last_sent_request.retain(|_, req_time| *req_time > time.elapsed_seconds() - 10.0);
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct AvatarSnapshots {
+    pub face256: String,
+    pub body: String,
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct AvatarEmote {
+    pub slot: u32,
+    pub urn: String,
+}
+
+#[derive(Serialize, Deserialize, Copy, Clone)]
+pub struct AvatarColor {
+    pub color: Color3,
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct AvatarWireFormat {
+    pub name: Option<String>,
+    #[serde(rename = "bodyShape")]
+    pub body_shape: Option<String>,
+    pub eyes: Option<AvatarColor>,
+    pub hair: Option<AvatarColor>,
+    pub skin: Option<AvatarColor>,
+    pub wearables: Vec<String>,
+    pub emotes: Option<Vec<AvatarEmote>>,
+    pub snapshots: Option<AvatarSnapshots>,
 }
 
 #[derive(Serialize, Deserialize, Clone)]
